Create a dynamic method to translate enum collections
For a field named "status" we now have a "Foo.statuses_i18n" method,
which is similar to the existing "foo_instance.status_i18n".

Refactored the code to keep everything understandable with the new
changes. It now uses "instance_eval" and "class_eval" with string arguments,
which is faster than with blocks. Helper methods were placed in a distinct
module to avoid ActiveRecord::Base namespace pollution.